Switch to new virtual floppy image (.ccvf) format
Previously, the polarity of the start, stop, and gap bits were the
opposite of real hardware, while the data bits were correct.  Now
everything matches the real hardware.

The previous checking concerning the CRT controller timing PROM should
have contained an update to roms.js, which is bundled in now.