invisible changes to the debugger
The debugger is disabled by default, and it still is. I've simply
rearranged a bit of code so that it is enabled via javascript, not css
changes.  Also, I added a clearing div after the floppy disk controls so
that if the debugger is enabled, the debugger buttons stay together.