using File API to save virtual disk images locally
For browsers which support it, the File API is used to save a blob
representation of the virtual disk image to local disk.

I've also disabled the 8.79 ROM option (in the code, just invisible in
presentation).

Fixed an assertion when a program attempted to write to an empty disk
drive.

I've cleaned up a few stylistic code nits while I was bumbling around.