code tweak re: boot ROM; no functional change
Rather than having deeply buried hide() and show() calls to
disable/enable ROM selection and the debugger interface, make two
controlling vars which are prominent in the file intro.  Also, there is
no need to reset the machine when setting the rom version if it matches
the existing ROM version.