Fix bugs in the 8080 core
Fixed bugs related to handling of the carry flags in SUB and SBB, and
handling of aux carry in SUB, SBB, DAA.  Added a hidden feature to
turbo-boost the CPU (though unclean).  Modified a hidden feature in the
key-at-a-time autotyper mode to delay a few CRT field times after each
line feed to allow the application to take care of business.